Changed quadratic b-spline example to use a per-element dof map. CompositeEvaluator.importMappedField now requires an explicit target domain (consistency NOT checked at the moment).

+++ trunk/src/fieldml/io/ReflectiveWalker.java
@@ -29,7 +29,7 @@ public class ReflectiveWalker
             }
             else if( o2.getClass() == Integer.class )
             {
-                handler.onIntListElement( (Integer)o2 );
+                handler.onIntListElement( o2 );
             }
             else if( f.isAnnotationPresent( SerializationAsString.class ) )
             {
